+++ config/tests.php
@@ -16,11 +16,17 @@ use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithClassAttri
 use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithClassAttributeWithoutExplicitProperty;
 use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithConstructorWithClassAttribute;
 use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithConstructorWithoutClassAttribute;
+use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithExtraArguments;
 use Rekalogika\Mapper\Tests\Fixtures\PropertyMapper\PropertyMapperWithoutClassAttribute;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
+
+    $services->defaults()
+        ->autowire()
+        ->autoconfigure()
+        ->public();
 
     // add test aliases
     $serviceIds = TestKernel::getServiceIds();
@@ -29,28 +35,10 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         $services->alias('test.' . $serviceId, $serviceId)->public();
     };
 
-    $services->set(PropertyMapperWithoutClassAttribute::class)
-        ->autowire()
-        ->autoconfigure()
-        ->public();
-
-    $services->set(PropertyMapperWithClassAttribute::class)
-        ->autowire()
-        ->autoconfigure()
-        ->public();
-
-    $services->set(PropertyMapperWithConstructorWithoutClassAttribute::class)
-        ->autowire()
-        ->autoconfigure()
-        ->public();
-
-    $services->set(PropertyMapperWithConstructorWithClassAttribute::class)
-        ->autowire()
-        ->autoconfigure()
-        ->public();
-
-    $services->set(PropertyMapperWithClassAttributeWithoutExplicitProperty::class)
-        ->autowire()
-        ->autoconfigure()
-        ->public();
+    $services->set(PropertyMapperWithoutClassAttribute::class);
+    $services->set(PropertyMapperWithClassAttribute::class);
+    $services->set(PropertyMapperWithConstructorWithoutClassAttribute::class);
+    $services->set(PropertyMapperWithConstructorWithClassAttribute::class);
+    $services->set(PropertyMapperWithClassAttributeWithoutExplicitProperty::class);
+    $services->set(PropertyMapperWithExtraArguments::class);
 };