+++ src/DependencyInjection/PropertyMapperPass.php
@@ -13,6 +13,10 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\DependencyInjection;
 
+use Rekalogika\Mapper\Context\Context;
+use Rekalogika\Mapper\Exception\InvalidArgumentException;
+use Rekalogika\Mapper\MainTransformer\MainTransformerInterface;
+use Rekalogika\Mapper\PropertyMapper\Contracts\PropertyMapperServicePointer;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -23,9 +27,16 @@ final class PropertyMapperPass implements CompilerPassInterface
         $propertyMapperResolver = $container
             ->getDefinition('rekalogika.mapper.property_mapper.resolver');
 
-        foreach ($container->findTaggedServiceIds('rekalogika.mapper.property_mapper') as $serviceId => $tags) {
+        $taggedServices = $container->findTaggedServiceIds('rekalogika.mapper.property_mapper');
+
+        foreach ($taggedServices as $serviceId => $tags) {
+            $serviceDefinition = $container->getDefinition($serviceId);
+            $serviceClass = $serviceDefinition->getClass() ?? throw new InvalidArgumentException('Class is required');
+
             /** @var array<string,string> $tag */
             foreach ($tags as $tag) {
+                $method = $tag['method'] ?? throw new InvalidArgumentException('Method is required');
+
                 $propertyMapperResolver->addMethodCall(
                     'addPropertyMapper',
                     [
@@ -33,10 +44,56 @@ final class PropertyMapperPass implements CompilerPassInterface
                         $tag['targetClass'],
                         $tag['property'],
                         $serviceId,
-                        $tag['method'],
+                        $method,
+                        self::getExtraArguments($serviceClass, $method),
                     ]
                 );
             }
         }
+    }
+
+    /**
+     * @param class-string $serviceClass
+     * @return array<int,PropertyMapperServicePointer::ARGUMENT_*>
+     */
+    private static function getExtraArguments(
+        string $serviceClass,
+        string $method
+    ): array {
+        $reflectionClass = new \ReflectionClass($serviceClass);
+        $parameters = $reflectionClass->getMethod($method)->getParameters();
+        // remove first element, which is always the source class
+        array_shift($parameters);
+
+        $extraArguments = [];
+
+        foreach ($parameters as $parameter) {
+            $type = $parameter->getType();
+
+            if (!$type instanceof \ReflectionNamedType) {
+                throw new InvalidArgumentException(
+                    sprintf(
+                        'Extra arguments for property mapper "%s" in class "%s" must be type hinted.',
+                        $method,
+                        $serviceClass,
+                    )
+                );
+            }
+
+            $extraArguments[] = match ($type->getName()) {
+                Context::class => PropertyMapperServicePointer::ARGUMENT_CONTEXT,
+                MainTransformerInterface::class => PropertyMapperServicePointer::ARGUMENT_MAIN_TRANSFORMER,
+                default => throw new InvalidArgumentException(
+                    sprintf(
+                        'Extra argument with type "%s" for property mapper "%s" in class "%s" is unsupported.',
+                        $type->getName(),
+                        $method,
+                        $serviceClass,
+                    )
+                )
+            };
+        }
+
+        return $extraArguments;
     }
 }