+++ tests/Common/TestKernel.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Tests\Common;
 
+use Rekalogika\Mapper\MapperInterface;
 use Rekalogika\Mapper\Mapping\MappingFactoryInterface;
 use Rekalogika\Mapper\RekalogikaMapperBundle;
 use Rekalogika\Mapper\Transformer\ArrayToObjectTransformer;
@@ -69,6 +70,8 @@ class TestKernel extends Kernel
      */
     public static function getServiceIds(): iterable
     {
+        yield MapperInterface::class;
+
         yield 'rekalogika.mapper.property_info';
         yield 'rekalogika.mapper.cache.property_info';
         yield 'rekalogika.mapper.property_info.cache';